requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
pydantic>=1.8.0
python-dotenv>=0.19.0
redis>=5.2.1
pydantic-settings>=2.9.1
pymysql>=1.1.1
jose>=1.0.0
python-jose>=3.4.0
requests>=2.32.3
pytz>=2025.2
websockets>=15.0.1
langgraph>=0.4.7
langchain_openai>=0.3.18